objectbuf/json: fix serialization crashes and harden decode
Crash root cause (heap corruption via unbounded Lua-stack growth in C
recursion, no api stack checks in release builds):
- objectbuf packer_table: add MAX_PACK_DEPTH cap + lua_checkstack per level,
  with balanced CTX.depth on every exit path.
- json encode/decode: same guard. parse_array/parse_object and
  encode_table now bound nesting by JSON_MAX_DEPTH (1000) and lua_checkstack
  each level. decode consumes untrusted input, so deep JSON was a remote DoS.

objectbuf data correctness / robustness:
- table encode: only treat genuine integer keys as array indices; a numeric
  string key like "1" was coerced by lua_tonumber and the pair silently dropped.
- encode(false) returned an empty string (lua_pushstring uses strlen on
  "\x00"); emit an explicit 1-byte string.
- validate symbol table shape on both encode and decode (index is a number
  >= TRUE_INDEX, map is a table) instead of blind rawget on possibly-nil.
- decode: reject empty input instead of silently returning boolean false.
- decode: bound `table` and array-part counts by remaining input bytes before
  pre-allocating, stopping a tiny malformed header from forcing billions of
  lua_newtable calls; make the array loop counter uint32 (no signed overflow).

stream_ffi ffi_stream_get_u30: cast (b & 127) to uint32 before shifting;
(b & 127) << shift overflowed signed int at shift 28 (undefined behavior).

function encoding semantics are unchanged.

+++ src/stream_ffi.c
@@ -40,7 +40,10 @@ bool ffi_stream_get_u30(BYTEARRAY *ba, uint32_t *result) {
         if (!bytearray_read8(ba, &b)) {
             return false;
         }
-        value |= ((b & 127) << shift);
+        // Cast before shifting: (b & 127) is an int, so `<< shift` with shift up
+        // to 28 would overflow signed int (undefined behavior). This varint is
+        // LEB128 over 32 bits (max 5 bytes); shift > 30 stops after the 5th byte.
+        value |= ((uint32_t)(b & 127)) << shift;
         shift += 7;
 
         if ((b & 128) == 0 || shift > 30) {
